Fjernar siste rest av at postnummer var int

+++ src/main/kotlin/no/roedt/innloggaBruker/InnloggaBrukerService.kt
@@ -29,7 +29,7 @@ class InnloggaBrukerService(
         etternavn = etternavn,
         telefonnummer = telefonnummer,
         email = email,
-        postnummer = postnummer.Postnummer.toInt(),
+        postnummer = postnummer.Postnummer,
         fylke = fylke,
         lokallag = lokallag,
         rolle = RingesentralenGroupID.from(groupID()).roller,
